рефакторинг по комментариям предыдущего задания
