+++ src/data.js
@@ -1,0 +1,4 @@
+export const KEY = {
+  ESC: 'Esc',
+  ESCAPE: 'Escape',
+};
