+++ src/main.js
@@ -1,25 +1,21 @@
 import ProfileView from './view/profile.js';
-import MenuListView from './view/menu.js';
+import MenuView from './view/menu.js';
 import ShowMoreButtonView from './view/show-more-button.js';
 import StatisticsView from './view/statistics.js';
 import SortView from './view/sort.js';
 import {generateFilms} from './mock/film.js';
-import {renderElement, RenderPosition} from './utils.js';
-import {generateFilter} from './mock/filters.js';
+import {renderElement, RenderPosition} from './utils-common.js';
+import {generateFilter} from './view/menu.js';
 import FilmsContainerView from './view/film-container.js';
 import FilmsListView from './view/film-list.js';
 // import FilmListExtraView from './view/film-list-extra.js';
 import FilmCardView from './view/film-card.js';
 import FilmDetailsView from './view/film-details.js';
+import {KEY} from './data.js';
 
 const FILM_CARDS_COUNT = 18;
 const FILM_CARDS_PER_STEP = 5;
 // const EXTRA_CARDS = 2;
-
-const Key = {
-  ESC: 'Esc',
-  ESCAPE: 'Escape',
-};
 
 const allFilms = generateFilms(FILM_CARDS_COUNT);
 const filters = generateFilter(allFilms);
@@ -30,7 +26,7 @@ const footerElement = document.querySelector('.footer');
 const bodyElement = document.querySelector('body');
 
 renderElement(headerElement, new ProfileView().getElement(), RenderPosition.BEFOREEND);
-renderElement(mainElement, new MenuListView(filters).getElement(), RenderPosition.BEFOREEND);
+renderElement(mainElement, new MenuView(filters).getElement(), RenderPosition.BEFOREEND);
 renderElement(mainElement, new SortView().getElement(), RenderPosition.BEFOREEND);
 
 const filmsContainer = new FilmsContainerView();
@@ -57,7 +53,7 @@ const renderFilmCards = (container, film) => {
     popup.getElement().querySelector('.film-details__close-btn').addEventListener('click', removePopup);
 
     const onEscKeyDown = (evt) => {
-      if (evt.key === Key.ESCAPE || evt.key === Key.ESC) {
+      if (evt.key === KEY.ESCAPE || evt.key === KEY.ESC) {
         evt.preventDefault();
         removePopup();
         document.removeEventListener('keydown', onEscKeyDown);
