+++ src/mock/film.js
@@ -1,5 +1,5 @@
-import { generateComments } from './comments.js';
-import {getRandomFloat, getRandomInteger } from '../utils.js';
+import {generateComments} from './comments.js';
+import {getRandomFloat, getRandomInteger } from './utils-for-mock.js';
 import dayjs from 'dayjs';
 
 const GENRES = [
