+++ src/utils-common.js
@@ -1,18 +1,3 @@
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomFloat = (min, max, decimalPoints = 1) => {
-  if (max > min && min >= 0) {
-    const randomNumber = Math.random() * (max - min + 1) + min;
-    return randomNumber.toFixed(decimalPoints);
-  }
-  return undefined;
-};
-
 const RenderPosition = {
   AFTEREND: 'afterend',
   BEFOREEND: 'beforeend',
@@ -40,4 +25,4 @@ const createElement = (template) => {
   return newElement.firstElementChild;
 };
 
-export {getRandomInteger, getRandomFloat, renderTemplate, renderElement, createElement, RenderPosition};
+export {renderTemplate, renderElement, createElement, RenderPosition};
