+++ src/view/profile.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils-common.js';
 
 const profileTemplate = () => (
   `<section class="header__profile profile">
