+++ src/view/sort.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils-common.js';
 
 const sortTemplate = () => (
   `<ul class="sort">
